requirements.txt
fastapi
uvicorn[standard]
gunicorn
requests
python-dotenv
supabase
